movements.models: Prices are decimal, not integer!

+++ apps/movements/models.py
@@ -126,7 +126,7 @@ class PurchaseOrderItem(models.Model):
     purchase_order = models.ForeignKey(PurchaseOrder, verbose_name=_(u'purchase order'))
     item_template = models.ForeignKey(ItemTemplate, verbose_name=_(u'item template'))
     status = models.ForeignKey(PurchaseRequestStatus, null=True, blank=True, verbose_name=_(u'status'))
-    agreed_price = models.PositiveIntegerField(null=True, blank=True, verbose_name=_(u'agreed price'))
+    agreed_price = models.DecimalField(max_digits=12, decimal_places=4, null=True, blank=True, verbose_name=_(u'agreed price'))
     active = models.BooleanField(default=True, verbose_name=_(u'active'))
     status = models.ForeignKey(PurchaseOrderItemStatus, null=True, blank=True, verbose_name=_(u'status'))
     qty = models.PositiveIntegerField(verbose_name=_(u'quantity'))
